AGS: Script API: added Object.AnimationVolume, like in Character
Missed to add this in 3.6.0 (thought it is not necessary for some reason...)
From upstream 480a8de9d3c5dc36eaaa3e4be94a1edf00445d68

+++ engines/ags/engine/ac/character_extras.cpp
@@ -44,8 +44,8 @@ void CharacterExtras::ReadFromSavegame(Stream *in, int save_ver) {
 	animwait = in->ReadInt16();
 	if (save_ver >= 2) // expanded at ver 2
 	{
-		anim_volume = in->ReadInt8();
-		cur_anim_volume = in->ReadInt8();
+		anim_volume = static_cast<uint8_t>(in->ReadInt8());
+		cur_anim_volume = static_cast<uint8_t>(in->ReadInt8());
 		in->ReadInt8(); // reserved to fill int32
 		in->ReadInt8();
 	}
